requirements.txt
anyio==4.1.0
appnope==0.1.3
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
comm==0.1.4
configparser==6.0.0
contourpy==1.1.0
cycler==0.11.0
db-dtypes==1.1.1
debugpy==1.6.7.post1
decorator==5.1.1
exceptiongroup==1.2.0
executing==1.2.0
fonttools==4.42.1
google-api-core==2.11.1
google-api-python-client==2.97.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-cloud-bigquery==3.13.0
google-cloud-bigquery-storage==2.22.0
google-cloud-core==2.3.3
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
google-search-results==2.4.2
google-serp-api==1.0.4
googleapis-common-protos==1.60.0
grpcio==1.57.0
grpcio-status==1.57.0
h11==0.14.0
httpcore==1.0.2
httplib2==0.22.0
httpx==0.25.2
idna==3.4
ipykernel==6.25.1
ipython==8.14.0
jedi==0.19.0
jupyter_client==8.3.0
jupyter_core==5.3.1
kiwisolver==1.4.5
matplotlib==3.7.2
matplotlib-inline==0.1.6
nest-asyncio==1.5.7
numpy==1.25.2
oauthlib==3.2.2
outcome==1.3.0.post0
packaging==23.1
pandas==2.0.3
pandas-gbq==0.19.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.10.0
prompt-toolkit==3.0.39
proto-plus==1.22.3
protobuf==4.24.1
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydata-google-auth==1.8.2
Pygments==2.16.1
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-telegram-bot==20.7
pytz==2023.3
pyzmq==25.1.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
seaborn==0.12.2
selenium==4.15.2
serpapi==0.1.5
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.23
stack-data==0.6.2
telegram==0.0.1
tornado==6.3.3
traitlets==5.9.0
trio==0.23.1
trio-websocket==0.11.1
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
wcwidth==0.2.6
wordcloud==1.9.2
wsproto==1.2.0
